// frontend/src/api/models/LabTestResultResponse.ts
export interface LabTestResultResponse {
    /**
     * 
     * @type {string}
     * @memberof LabTestResultResponse
     */
    id?: string;
    /**
     * 
     * @type {LabTestResponse}
     * @memberof LabTestResultResponse
     */
    labTest?: LabTestResponse;
    /**
     * 
     * @type {string}
     * @memberof LabTestResultResponse
     */
    status?: LabTestResultResponseStatusEnum;
    /**
     * 
     * @type {string}
     * @memberof LabTestResultResponse
     */
    resultValue?: string;
    /**
     * 
     * @type {string}
     * @memberof LabTestResultResponse
     */
    unit?: string;
    /**
     * 
     * @type {string}
     * @memberof LabTestResultResponse
     */
    referenceRange?: string;
    /**
     * 
     * @type {string}
     * @memberof LabTestResultResponse
     */
    remark?: string;
    /**
     * 
     * @type {StaffResponse}
     * @memberof LabTestResultResponse
     */
    verifiedBy?: StaffResponse;
    /**
     * 
     * @type {Date}
     * @memberof LabTestResultResponse
     */
    verifiedAt?: Date;
}


/**
 * @export
 */
export const LabTestResultResponseStatusEnum = {
    Pending: 'PENDING',
    InProgress: 'IN_PROGRESS',
    Completed: 'COMPLETED',
    Verified: 'VERIFIED',
    Cancelled: 'CANCELLED'
} as const;
export type LabTestResultResponseStatusEnum = typeof LabTestResultResponseStatusEnum[keyof typeof LabTestResultResponseStatusEnum];